+++ src/javascript/common/keyPressed/key_pressed.js
@@ -1,7 +1,7 @@
 // @flow
 /* eslint-disable no-console,no-use-before-define */
 
-import type { KeyPressedEvent, KeyPressedOptions } from '../../type/key_pressed';
+import type { KeyPressedEvent, KeyPressedOptions } from '../../type/key_pressed-type';
 
 const keyPressedClass = (defaultOptions: KeyPressedOptions = {}) => {
   const prop: WeakMap<KeyPressed, KpUtils> = new WeakMap();
