+++ src/javascript/common/tiles/board-tiles.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type { Tile } from '../../type/tile';
+import type { Tile } from '../../type/tile-type';
 
 export const boardTiles: Tile = {
   x: 340,
