+++ src/javascript/common/tiles/corner-tiles.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type { Tile } from '../../type/tile';
+import type { Tile } from '../../type/tile-type';
 
 export const topLeft: Tile = {
   x: 203,
