+++ src/javascript/common/tiles/stat-tiles.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type { Tile } from '../../type/tile';
+import type { Tile } from '../../type/tile-type';
 
 export const zeroCardTile: Tile = {
   x: 356,
