+++ src/javascript/constructor/animations-constructor.js
