+++ src/javascript/engine/key-handler.js
@@ -1,14 +1,14 @@
 // @flow
 
-import type { GridPosition } from '../type/canvas';
+import type { GridPosition } from '../type/canvas-type';
 import {
   battlegroundPositions,
   isValidBattlegroundPosition,
 } from '../common/positions/battleground-positions';
 import { DOWN, LEFT, RIGHT, UP } from '../common/variables';
 import { isValidPlayerHandPosition, playerHandPositions } from '../common/positions/player-side-positions';
-import Monster from '../constructor/card/monster';
-import Stone from '../constructor/card/stone';
+import Stone from '../constructor/card/stone-constructor';
+import Monster from '../constructor/card/monster-constructor';
 
 export const isBattlegroundGridPositionAvailable = (
   listOfCards: Array<?Stone | ?Monster>,
