+++ src/javascript/type/canvas-type.js
