+++ src/javascript/type/key_pressed-type.js
