+++ src/javascript/type/stat-type.js
