+++ src/javascript/type/store-type.js
